OCT-111: moved FindOneAttributeByCodeQuery
