+++ components/catalogs/back/src/Application/Persistence/Attribute/FindOneAttributeByCodeQueryInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Application\Persistence;
+namespace Akeneo\Catalogs\Application\Persistence\Attribute;
 
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
