+++ components/catalogs/back/src/Infrastructure/Symfony/Resources/config/services.yml
@@ -88,7 +88,7 @@ services:
         arguments:
             - '@pim_enrich.repository.attribute_option.search'
 
-    Akeneo\Catalogs\Infrastructure\Persistence\FindOneAttributeByCodeQuery:
+    Akeneo\Catalogs\Infrastructure\Persistence\Attribute\FindOneAttributeByCodeQuery:
         arguments:
             $repository: '@pim_catalog.repository.attribute'
