+++ components/catalogs/back/src/Infrastructure/Validation/ProductSelection/AttributeCriterionContainsValidScopeValidator.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Akeneo\Catalogs\Infrastructure\Validation\ProductSelection;
 
+use Akeneo\Catalogs\Application\Persistence\Attribute\FindOneAttributeByCodeQueryInterface;
 use Akeneo\Catalogs\Application\Persistence\Channel\GetChannelQueryInterface;
-use Akeneo\Catalogs\Application\Persistence\FindOneAttributeByCodeQueryInterface;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
 use Symfony\Component\Validator\Exception\UnexpectedTypeException;
